requirements.txt
python-binance
requests==2.32.4
python-dotenv==1.1.1
Flask==3.1.1